refactor(visualization): move code to separate file
